#44 - searchType 폴더명 변경
- type에서 constant로 변경함
