--- src/main/java/com/fastcampus/mvcboardproject/domain/type/SearchType.java
@@ -1,18 +1,0 @@
-package com.fastcampus.mvcboardproject.domain.type;
-
-import lombok.Getter;
-
-public enum SearchType {
-    TITLE("제목"),
-    CONTENT("본문"),
-    ID("유저 ID"),
-    NICKNAME("닉네임"),
-    HASHTAG("해시태그");
-
-    @Getter
-    private final String description;
-
-    SearchType(String description) {
-        this.description = description;
-    }
-}
